# packages/origine2/src/config/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/manageGame/createGame": {
      "post": {
        "operationId": "ManageGameController_createGame",
        "summary": "Create a new game",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Game creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game creation result."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/openGameDict/{gameName}": {
      "get": {
        "operationId": "ManageGameController_openGameDict",
        "summary": "Open Game Dictionary",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "description": "Name of the game.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opens the dictionary for a specified game."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
...
    "/api/manageGame/openGameAssetsDict/{gameName}": {
      "get": {
        "operationId": "ManageGameController_openGameAssetsDict",
        "summary": "Open Game Assets Dictionary",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "description": "Name of the game.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subFolder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opens the assets dictionary for a specified game."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/ejectGameAsWeb/{gameName}": {
      "get": {
        "operationId": "ManageGameController_ejectGameAsWeb",
        "summary": "Eject Game As Web App",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "description": "Name of the game to be exported as web app.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exports the specified game as a web app."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/ejectGameAsExe/{gameName}": {
      "get": {
        "operationId": "ManageGameController_ejectGameAsExe",
        "summary": "Eject Game As EXE",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "description": "Name of the game to be exported as EXE.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exports the specified game as an EXE (Windows Electron App)."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/ejectGameAsAndroid/{gameName}": {
      "get": {
        "operationId": "ManageGameController_ejectGameAsAndroid",
        "summary": "Eject Game As Android App",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "description": "Name of the game to be exported as an Android app.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exports the specified game as an Android app."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/readGameAssets/{readDirPath}": {
      "get": {
        "operationId": "ManageGameController_readGameAssets",
        "summary": "Read Game Assets",
        "parameters": [
          {
            "name": "readDirPath",
            "required": true,
            "in": "path",
            "description": "Path of the game directory to read assets from, including subdirectories.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieve the assets of the specified game directory."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/editFileName": {
      "post": {
        "operationId": "ManageGameController_editFileName",
        "summary": "Edit File Name",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "File renaming data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditFileNameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully renamed the file."
          },
          "400": {
            "description": "Failed to rename the file."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/deleteFile": {
      "post": {
        "operationId": "ManageGameController_deleteFile",
        "summary": "Delete File",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "File deletion data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deleted the file."
          },
          "400": {
            "description": "Failed to delete the file."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/createNewScene": {
      "post": {
        "operationId": "ManageGameController_createNewScene",
        "summary": "Create a New Scene",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Scene creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewSceneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created the scene."
          },
          "400": {
            "description": "Failed to create the scene or scene already exists."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/editScene": {
      "post": {
        "operationId": "ManageGameController_editScene",
        "summary": "Edit Scene",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSceneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scene edited successfully."
          },
          "400": {
            "description": "Failed to edit the scene."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/editTextFile": {
      "post": {
        "operationId": "ManageGameController_editTextFile",
        "summary": "Edit TextFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTextFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File edited successfully."
          },
          "400": {
            "description": "Failed to edit the file."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/getGameConfig/{gameName}": {
      "get": {
        "operationId": "ManageGameController_getGameConfig",
        "summary": "Get Game Configuration",
        "parameters": [
          {
            "name": "gameName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned game configuration."
          },
          "400": {
            "description": "Failed to get the game configuration."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/setGameConfig": {
      "post": {
        "operationId": "ManageGameController_setGameConfig",
        "summary": "Set Game Configuration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game configuration set successfully."
          },
          "400": {
            "description": "Failed to set the game configuration."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/uploadFiles": {
      "post": {
        "operationId": "ManageGameController_uploadFiles",
        "summary": "Upload Files",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Files uploaded successfully."
          },
          "400": {
            "description": "Failed to upload files."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/mkdir": {
      "post": {
        "operationId": "ManageGameController_mkDir",
        "summary": "Create Directory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MkDirDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Directory created successfully."
          },
          "400": {
            "description": "Failed to create directory."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/delete": {
      "post": {
        "operationId": "ManageGameController_delete",
        "summary": "Delete File or Directory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deleted the file or directory."
          },
          "400": {
            "description": "Failed to delete the file or directory."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/rename": {
      "post": {
        "operationId": "ManageGameController_rename",
        "summary": "Rename File or Directory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully renamed the file or directory."
          },
          "400": {
            "description": "Failed to rename the file or directory."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
    "/api/manageGame/getIcons/{gameDir}": {
      "get": {
        "operationId": "ManageGameController_getIcons",
        "summary": "Get Game Icons",
        "parameters": [
          {
            "name": "gameDir",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned game icons.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IconsDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to get the game icons."
          }
        },
        "tags": [
          "Manage Game"
        ]
      }
    },
...
  },
  "info": {
    "title": "WebGAL Terre API",
    "description": "API Refrence of WebGAL Terre Editor",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UploadFilesDto": {
        "type": "object",
        "properties": {
          "targetDirectory": {
            "type": "string",
            "description": "Target directory for the uploaded files"
          }
        },
        "required": [
          "targetDirectory"
        ]
      },
...
      "EditTextFileDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "The path of textfile"
          },
          "textFile": {
            "type": "string",
            "description": "Text data content"
          }
        },
        "required": [
          "path",
          "textFile"
        ]
      },
...
      "CreateGameDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "The name of the game to be created"
          },
          "gameDir": {
            "type": "string",
            "description": "The dir of the game to be created"
          },
          "derivative": {
            "type": "string",
            "description": "The name of the derivative to be used"
          },
          "templateDir": {
            "type": "string",
            "description": "The dir of the template to be applied"
          },
          "ignoreTemplate": {
            "type": "boolean",
            "description": "Whether to ignore the template when creating the game"
          }
        },
        "required": [
          "gameName",
          "gameDir"
        ]
      },
      "EditFileNameDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "The path to the file to be renamed"
          },
          "newName": {
            "type": "string",
            "description": "The new name for the file"
          }
        },
        "required": [
          "path",
          "newName"
        ]
      },
      "DeleteFileDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "The path to the file to be deleted"
          }
        },
        "required": [
          "path"
        ]
      },
      "CreateNewSceneDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "The name of the game"
          },
          "sceneName": {
            "type": "string",
            "description": "The name of the scene"
          }
        },
        "required": [
          "gameName",
          "sceneName"
        ]
      },
      "EditSceneDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "The name of the game"
          },
          "sceneName": {
            "type": "string",
            "description": "The name of the scene"
          },
          "sceneData": {
            "type": "string",
            "description": "Scene data content",
            "format": "{ value: string }"
          }
        },
        "required": [
          "gameName",
          "sceneName",
          "sceneData"
        ]
      },
      "GameConfigDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "The name of the game"
          },
          "newConfig": {
            "type": "string",
            "description": "New game configuration"
          }
        },
        "required": [
          "gameName",
          "newConfig"
        ]
      },
      "MkDirDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "description": "The source path where the directory will be created"
          },
          "name": {
            "type": "string",
            "description": "Name for the new directory"
          }
        },
        "required": [
          "source",
          "name"
        ]
      },
      "DeleteDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "The source path of the file or directory to be deleted"
          }
        },
        "required": [
          "gameName"
        ]
      },
      "RenameDto": {
        "type": "object",
        "properties": {
          "gameName": {
            "type": "string",
            "description": "Old name for renaming the game"
          },
          "newName": {
            "type": "string",
            "description": "New name for renaming the game"
          }
        },
        "required": [
          "gameName",
          "newName"
        ]
      },
      "IconsDto": {
        "type": "object",
        "properties": {
          "platforms": {
            "description": "The icons of the game",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "platforms"
        ]
      },
...
    }
  }
}